# test: Enable type checking of tests