--- a/app/scripts/controllers/encryption-public-key.test.ts
+++ b/app/scripts/controllers/encryption-public-key.test.ts
@@ -214,6 +214,7 @@ describe('EncryptionPublicKeyController', () => {
   });
 
   describe('newRequestEncryptionPublicKey', () => {
+    // @ts-expect-error This function is missing from the Mocha type definitions
     it.each([
       ['Ledger', KeyringType.ledger],
       ['Trezor', KeyringType.trezor],
@@ -221,7 +222,10 @@ describe('EncryptionPublicKeyController', () => {
       ['QR hardware', KeyringType.qr],
     ])(
       'throws if keyring is not supported',
-      async (keyringName, keyringType) => {
+      async (
+        keyringName: string,
+        keyringType: (typeof KeyringType)[keyof typeof KeyringType],
+      ) => {
         getAccountKeyringTypeMock.mockResolvedValueOnce(keyringType);
 
         await expect(