--- a/app/scripts/lib/AccountIdentitiesPetnamesBridge.test.ts
+++ b/app/scripts/lib/AccountIdentitiesPetnamesBridge.test.ts
@@ -205,6 +205,7 @@ describe('AccountIdentitiesPetnamesBridge', () => {
   });
 
   describe('shouldSyncPetname', () => {
+    // @ts-expect-error This is missing from the Mocha type definitions
     it.each([
       {
         origin: NameOrigin.ACCOUNT_IDENTITY,
@@ -216,7 +217,13 @@ describe('AccountIdentitiesPetnamesBridge', () => {
       },
     ])(
       'returns $expectedReturn if origin is $origin',
-      ({ origin, expectedReturn }) => {
+      ({
+        origin,
+        expectedReturn,
+      }: {
+        origin: NameOrigin;
+        expectedReturn: boolean;
+      }) => {
         class TestBridge extends AccountIdentitiesPetnamesBridge {
           public shouldSyncPetname(entry: PetnameEntry): boolean {
             return super.shouldSyncPetname(entry);