--- a/app/scripts/lib/snap-keyring/utils/isBlockedUrl.test.ts
+++ b/app/scripts/lib/snap-keyring/utils/isBlockedUrl.test.ts
@@ -27,6 +27,7 @@ describe('isBlockedUrl', () => {
     },
   });
 
+  // @ts-expect-error This is missing from the Mocha type definitions
   it.each([
     ['http://metamask.io', false],
     ['https://metamask.io', false],