--- a/app/scripts/migrations/095.test.ts
+++ b/app/scripts/migrations/095.test.ts
@@ -156,12 +156,13 @@ describe('migration #95', () => {
       });
     });
 
+    // @ts-expect-error This is missing from the Mocha type definitions
     it.each([
       ['undefined', undefined],
       ['empty', {}],
     ])(
       'does nothing if incoming transactions %s',
-      async (_title, incomingTransactions) => {
+      async (_title: string, incomingTransactions: unknown) => {
         const oldData = {
           some: 'data',
           IncomingTransactionsController: {