--- a/test/e2e/accounts/remove-account-snap.spec.ts
+++ b/test/e2e/accounts/remove-account-snap.spec.ts
@@ -1,4 +1,4 @@
-const { strict: assert } = require('assert');
+import { strict as assert } from 'assert';
 import { Suite } from 'mocha';
 import FixtureBuilder from '../fixture-builder';
 import { WINDOW_TITLES, defaultGanacheOptions, withFixtures } from '../helpers';