--- a/test/e2e/flask/user-operations.spec.ts
+++ b/test/e2e/flask/user-operations.spec.ts
@@ -1,10 +1,7 @@
-import { TransactionParams } from '@metamask/transaction-controller';
 import {
   withFixtures,
   unlockWallet,
-  openDapp,
   switchToNotificationWindow,
-  DAPP_URL,
   WINDOW_TITLES,
   sendTransaction,
   convertETHToHexGwei,