--- a/test/e2e/mmi/helpers/dapps-helpers.ts
+++ b/test/e2e/mmi/helpers/dapps-helpers.ts
@@ -74,6 +74,6 @@ export async function callTestDappBtn(
   );
   return {
     dummyDApp,
-    signedTransactionTime
-  }
+    signedTransactionTime,
+  };
 }