--- a/test/e2e/mmi/pageObjects/mmi-main-page.ts
+++ b/test/e2e/mmi/pageObjects/mmi-main-page.ts
@@ -27,7 +27,9 @@ export class MMIMainPage {
 
   async finishOnboarding() {
     await this.page.getByRole('button', { name: /continue/iu }).click();
-    await this.page.getByRole('button', { name: /continue to wallet/iu }).click();
+    await this.page
+      .getByRole('button', { name: /continue to wallet/iu })
+      .click();
   }
 
   async checkLastTransactionStatus(status: string | RegExp) {