--- a/test/e2e/mmi/pageObjects/mmi-mainMenu-page.ts
+++ b/test/e2e/mmi/pageObjects/mmi-mainMenu-page.ts
@@ -89,15 +89,18 @@ export class MMIMainMenuPage {
   }
 
   async fillPassword() {
-    await this.page.getByTestId('unlock-password').fill(process.env.MMI_E2E_MMI_PASSWORD as string);
+    await this.page
+      .getByTestId('unlock-password')
+      .fill(process.env.MMI_E2E_MMI_PASSWORD as string);
     await this.page.getByRole('button', { name: /unlock/iu }).click();
   }
 
   async finishOnboarding() {
     await this.page.getByRole('button', { name: /continue/iu }).click();
-    await this.page.getByRole('button', { name: /continue to wallet/iu }).click();
+    await this.page
+      .getByRole('button', { name: /continue to wallet/iu })
+      .click();
   }
-
 
   async switchTestNetwork() {
     await this.page