--- a/test/e2e/mmi/specs/navigation.spec.ts
+++ b/test/e2e/mmi/specs/navigation.spec.ts
@@ -11,7 +11,6 @@ import { Auth0Page } from '../pageObjects/mmi-auth0-page';
 import { MMIMainPage } from '../pageObjects/mmi-main-page';
 
 const portfolio = `${process.env.MMI_E2E_MMI_DASHBOARD_URL}/portfolio`;
-const swap = `${process.env.MMI_E2E_MMI_DASHBOARD_URL}/swap`;
 const stake = `${process.env.MMI_E2E_MMI_DASHBOARD_URL}/stake`;
 const support = 'https://mmi-support.metamask.io/hc/en-us';
 const supportContactUs =