--- a/test/e2e/mmi/specs/transactions.spec.ts
+++ b/test/e2e/mmi/specs/transactions.spec.ts
@@ -115,12 +115,8 @@ test.describe('MMI send', () => {
     const client = new CustodianTestClient();
     await client.setup();
     const repeatTx = true;
-    const { mainPage, custodianTxId, secondCustodianTxId } = await sendTransaction(
-      page,
-      context,
-      client,
-      repeatTx,
-    );
+    const { mainPage, custodianTxId, secondCustodianTxId } =
+      await sendTransaction(page, context, client, repeatTx);
 
     // Sign and submit
     const statusName = await client.submitTransactionById(custodianTxId);
@@ -130,7 +126,6 @@ test.describe('MMI send', () => {
       await client.submitTransactionById(secondCustodianTxId);
     }
   });
-
 
   test('Send a transaction from one account to another and abort it from custody', async ({
     page,