--- a/test/e2e/tests/confirmations/helpers.ts
+++ b/test/e2e/tests/confirmations/helpers.ts
@@ -1,18 +1,20 @@
-import { strict as assert } from 'assert';
 import FixtureBuilder from '../../fixture-builder';
-import {
-  defaultGanacheOptions,
-  withFixtures,
-} from '../../helpers';
+import { defaultGanacheOptions, withFixtures } from '../../helpers';
 import { Driver } from '../../webdriver/driver';
 
-export async function scrollAndConfirmAndAssertConfirm (driver: Driver) {
-  const confirmButton = await driver.findElement({ xpath: '//button[@data-testid="confirm-footer-button"]'});
+export async function scrollAndConfirmAndAssertConfirm(driver: Driver) {
   await driver.clickElement('.confirm-scroll-to-bottom__button');
   await driver.clickElement('[data-testid="confirm-footer-button"]');
 }
 
-export function withRedesignConfirmationFixtures (title: string = '', testFunction: Function) {
+export function withRedesignConfirmationFixtures(
+  // Default params first is discouraged because it makes it hard to call the function without the
+  // optional parameters. But it doesn't apply here because we're always passing in a variable for
+  // title. It's optional because it's sometimes unset.
+  // eslint-disable-next-line @typescript-eslint/default-param-last
+  title: string = '',
+  testFunction: Parameters<typeof withFixtures>[1],
+) {
   return withFixtures(
     {
       dapp: true,