--- a/test/e2e/tests/confirmations/signatures/personal-sign.spec.ts
+++ b/test/e2e/tests/confirmations/signatures/personal-sign.spec.ts
@@ -12,12 +12,20 @@ import { Ganache } from '../../../seeder/ganache';
 import { Driver } from '../../../webdriver/driver';
 
 describe('Confirmation Signature - Personal Sign', function (this: Suite) {
-  if (!process.env.ENABLE_CONFIRMATION_REDESIGN) { return; }
+  if (!process.env.ENABLE_CONFIRMATION_REDESIGN) {
+    return;
+  }
 
   it('initiates and confirms', async function () {
     await withRedesignConfirmationFixtures(
       this.test?.fullTitle(),
-      async ({ driver, ganacheServer }: { driver: Driver, ganacheServer: Ganache }) => {
+      async ({
+        driver,
+        ganacheServer,
+      }: {
+        driver: Driver;
+        ganacheServer: Ganache;
+      }) => {
         const addresses = await ganacheServer.getAccounts();
         const publicAddress = addresses?.[0] as string;
 
@@ -31,7 +39,7 @@ describe('Confirmation Signature - Personal Sign', function (this: Suite) {
         await driver.clickElement('[data-testid="confirm-footer-button"]');
 
         await assertVerifiedPersonalMessage(driver, publicAddress);
-      }
+      },
     );
   });
 
@@ -44,7 +52,9 @@ describe('Confirmation Signature - Personal Sign', function (this: Suite) {
         await driver.clickElement('#personalSign');
         await switchToNotificationWindow(driver);
 
-        await driver.clickElement('[data-testid="confirm-footer-cancel-button"]');
+        await driver.clickElement(
+          '[data-testid="confirm-footer-cancel-button"]',
+        );
 
         await driver.waitUntilXWindowHandles(2);
         await driver.switchToWindowWithTitle(WINDOW_TITLES.TestDApp);
@@ -61,23 +71,32 @@ describe('Confirmation Signature - Personal Sign', function (this: Suite) {
 
 async function assertInfoValues(driver: Driver) {
   const origin = driver.findElement({ text: DAPP_HOST_ADDRESS });
-  const message = driver.findElement({ text: 'Example `personal_sign` message' });
+  const message = driver.findElement({
+    text: 'Example `personal_sign` message',
+  });
 
   assert.ok(await origin);
   assert.ok(await message);
 }
 
-async function assertVerifiedPersonalMessage(driver: Driver, publicAddress: string) {
+async function assertVerifiedPersonalMessage(
+  driver: Driver,
+  publicAddress: string,
+) {
   await driver.waitUntilXWindowHandles(2);
   await driver.switchToWindowWithTitle(WINDOW_TITLES.TestDApp);
   await driver.clickElement('#personalSignVerify');
 
-  const verifySigUtil = await driver.findElement('#personalSignVerifySigUtilResult');
+  const verifySigUtil = await driver.findElement(
+    '#personalSignVerifySigUtilResult',
+  );
   await driver.waitForSelector({
     css: '#personalSignVerifyECRecoverResult',
     text: publicAddress,
   });
-  const verifyECRecover = await driver.findElement('#personalSignVerifyECRecoverResult');
+  const verifyECRecover = await driver.findElement(
+    '#personalSignVerifyECRecoverResult',
+  );
 
   assert.equal(await verifySigUtil.getText(), publicAddress);
   assert.equal(await verifyECRecover.getText(), publicAddress);