… test/e2e/tests/confirmations/signatures/sign-typed-data-v3.spec.ts
… test/e2e/tests/confirmations/signatures/sign-typed-data-v4.spec.ts
… test/e2e/tests/confirmations/signatures/sign-typed-data.spec.ts
… test/e2e/tests/ppom/mocks/mock-server-json-rpc.ts
… test/e2e/tests/simulation-details/mock-request-buy-erc1155.ts
… test/e2e/tests/simulation-details/mock-request-buy-erc20.ts
… test/e2e/tests/simulation-details/mock-request-buy-erc721.ts
… test/e2e/tests/simulation-details/mock-request-error-insuffient-gas.ts
… test/e2e/tests/simulation-details/mock-request-error-malformed-transaction.ts
… test/e2e/tests/simulation-details/mock-request-no-changes.ts
… test/e2e/tests/simulation-details/mock-request-send-eth.ts
… test/e2e/tests/simulation-details/simulation-details.spec.ts
… test/e2e/tests/simulation-details/types.ts
… test/e2e/tests/swap-send/mocks/erc20-data.ts
… test/e2e/tests/swap-send/mocks/eth-data.ts
… test/e2e/tests/swap-send/swap-send-erc20.spec.ts
… test/e2e/tests/swap-send/swap-send-eth.spec.ts
… test/e2e/tests/swap-send/swap-send-test-utils.ts
… tsconfig.json
… ui/components/app/alert-system/utils.test.ts
… ui/components/app/name/name.test.tsx
… ui/ducks/locale/locale.test.ts
… ui/pages/confirmations/components/simulation-details/amount-pill.test.tsx
… ui/pages/confirmations/components/simulation-details/asset-pill.test.tsx
… ui/pages/confirmations/components/simulation-details/fiat-display.test.tsx
… ui/pages/confirmations/components/simulation-details/formatAmount.test.ts
… ui/pages/confirmations/components/simulation-details/sortBalanceChanges.test.ts
… ui/pages/confirmations/components/simulation-details/useSimulationMetrics.test.ts
… ui/pages/confirmations/hooks/alerts/utils.test.ts