@Gudahtt 2024-05-28
I did some performance testing, since it was brought up in Slack that we found that type checking the `test` directory was too slow in the past. Here's what I found when comparing with `develop`:

### Before

```
yarn lint  163.65s user 9.55s system 134% cpu 2:09.17 total
yarn lint  161.45s user 9.33s system 128% cpu 2:04.26 total
yarn lint  139.80s user 9.18s system 126% cpu 1:57.80 total
```

### After

```
yarn lint  146.09s user 9.57s system 128% cpu 2:01.45 total
yarn lint  146.07s user 9.03s system 130% cpu 1:58.85 total
yarn lint  170.71s user 9.84s system 133% cpu 2:14.78 total
yarn lint  145.43s user 9.18s system 129% cpu 1:59.71 total
```

Seeing some unexpected bimodal results there, not sure why. I had plenty of other applications open during these tests, but nothing significant, and no major difference that I'm aware of between each test. They seem roughly in the same ballpark though.